Fix rebase issues
